backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic
cryptography
pycardano
requests
httpx
